# src/server/src/resources/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the API contract for the server for the DMS LEAP application",
    "version": "1.0.0",
    "title": "LEAP DMS APIs",
    "termsOfService": ""
  },
...
  "paths": {
    "/apps": {
      "post": {
        "tags": [
          "App Store"
        ],
        "summary": "Register new app",
        "description": "Creates new app based on data passed in the request body",
        "operationId": "createApp",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/App"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apps"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "App Store"
        ],
        "summary": "Gets all apps",
        "description": "Fetch the list of the external apps",
        "operationId": "getallApps",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/apps/{id}": {
      "get": {
        "tags": [
          "App Store"
        ],
        "summary": "Find app by id",
        "description": "Returns app based on id",
        "operationId": "getAppById",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of app to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apps"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "App not found"
          }
        }
      },
      "put": {
        "tags": [
          "App Store"
        ],
        "summary": "Updates the app",
        "description": "updates app for any of the fields provided in request payload",
        "operationId": "updateAppById",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of app to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUpdateResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/users/{id}/apps": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get apps by userId",
        "description": "Get all apps for user based on userId",
        "operationId": "getAllAppsByUserId",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of app to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserApps"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://leapapi-dev.leap.dellmed.utexas.edu/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "App": {
        "type": "object",
        "description": "the external app.",
        "properties": {
          "appName": {
            "type": "string",
            "description": "The commercial name of the app",
            "example": "Study App"
          },
          "appLogo": {
            "type": "string",
            "example": "https://url_to_app_logo_location"
          },
          "appUrl": {
            "type": "string",
            "example": "https://url_to_app_url_location"
          },
          "shortDescription": {
            "type": "string",
            "description": "The short description about the app",
            "example": "Short description of app"
          },
          "longDescription": {
            "type": "string",
            "description": "The long description about the app",
            "example": "Long description of app"
          },
          "consentVersion": {
            "type": "string",
            "example": "v1"
          },
          "consentTermOfUse": {
            "type": "string",
            "example": "test"
          },
          "consentPrivacy": {
            "type": "string",
            "example": "test"
          },
          "consentInformation": {
            "type": "string",
            "example": "test"
          },
          "consentPolicy": {
            "type": "string",
            "example": "<b>test</b>"
          },
          "isSSOEnabled": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Apps": {
        "type": "object",
        "description": "the external app.",
        "properties": {
          "_id": {
            "type": "string",
            "example": "5ff53d85af962933a8b3a555"
          },
          "appName": {
            "type": "string",
            "description": "The commercial name of the app",
            "example": "Study App"
          },
          "appLogo": {
            "type": "string",
            "example": "https://url_to_app_logo_location"
          },
          "appUrl": {
            "type": "string",
            "example": "https://url_to_app_url_location"
          },
          "shortDescription": {
            "type": "string",
            "description": "The short description about the app",
            "example": "Short description of app"
          },
          "longDescription": {
            "type": "string",
            "description": "The long description about the app",
            "example": "Long description of app"
          },
          "consentVersion": {
            "type": "string",
            "example": "v1"
          },
          "consentTermOfUse": {
            "type": "string",
            "example": "test"
          },
          "consentPrivacy": {
            "type": "string",
            "example": "test"
          },
          "consentInformation": {
            "type": "string",
            "example": "test"
          },
          "consentPolicy": {
            "type": "string",
            "example": "<b>test</b>"
          },
          "isSSOEnabled": {
            "type": "boolean",
            "example": false
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-06T04:33:09.601+00:00"
          },
          "createdBy": {
            "type": "string",
            "example": "user name"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-06T04:33:09.601+00:00"
          },
          "updatedBy": {
            "type": "string",
            "example": "user name"
          },
          "notifications": {
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          },
          "medicalRecords": {
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AppUpdateRequest": {
        "type": "object",
        "description": "the external app.",
        "properties": {
          "appName": {
            "type": "string",
            "description": "The commercial name of the app",
            "example": "Study App"
          },
          "appLogo": {
            "type": "string",
            "example": "https://url_to_app_logo_location"
          },
          "appUrl": {
            "type": "string",
            "example": "https://url_to_app_url_location"
          },
          "shortDescription": {
            "type": "string",
            "description": "The short description about the app",
            "example": "Short description of app"
          },
          "longDescription": {
            "type": "string",
            "description": "The long description about the app",
            "example": "Updated long description of app"
          },
          "consentVersion": {
            "type": "string",
            "example": "v2"
          },
          "consentTermOfUse": {
            "type": "string",
            "example": "test"
          },
          "consentPrivacy": {
            "type": "string",
            "example": "test"
          },
          "consentInformation": {
            "type": "string",
            "example": "test"
          },
          "consentPolicy": {
            "type": "string",
            "example": "<b>test</b>"
          },
          "isSSOEnabled": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "AppUpdateResponse": {
        "type": "object",
        "description": "the external app.",
        "properties": {
          "_id": {
            "type": "string",
            "example": "5ff53d85af962933a8b3a555"
          },
          "appName": {
            "type": "string",
            "description": "The commercial name of the app",
            "example": "Study App"
          },
          "appLogo": {
            "type": "string",
            "example": "https://url_to_app_logo_location"
          },
          "appUrl": {
            "type": "string",
            "example": "https://url_to_app_url_location"
          },
          "shortDescription": {
            "type": "string",
            "description": "The short description about the app",
            "example": "Short description of app"
          },
          "longDescription": {
            "type": "string",
            "description": "The long description about the app",
            "example": "Updated long description of app"
          },
          "consentVersion": {
            "type": "string",
            "example": "v2"
          },
          "consentTermOfUse": {
            "type": "string",
            "example": "test"
          },
          "consentPrivacy": {
            "type": "string",
            "example": "test"
          },
          "consentInformation": {
            "type": "string",
            "example": "test"
          },
          "consentPolicy": {
            "type": "string",
            "example": "<b>test</b>"
          },
          "isSSOEnabled": {
            "type": "boolean",
            "example": false
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-06T04:33:09.601+00:00"
          },
          "createdBy": {
            "type": "string",
            "example": "user name"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-06T04:33:09.601+00:00"
          },
          "updatedBy": {
            "type": "string",
            "example": "user name"
          },
          "notifications": {
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          },
          "medicalRecords": {
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AppResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "appName": {
              "type": "string"
            },
            "appLogo": {
              "type": "string"
            },
            "appUrl": {
              "type": "string"
            },
            "longDescription": {
              "type": "string"
            },
            "shortDescription": {
              "type": "string"
            },
            "isActive": {
              "type": "boolean"
            }
          }
        },
        "example": [
          {
            "_id": "5ff53d85af962933a8b3a555",
            "appName": "Study App",
            "appLogo": "https://url_to_app_logo_location",
            "appUrl": "https://url_to_app_url_location",
            "longDescription": "Long description of app",
            "shortDescription": "Short description of app",
            "isActive": true
          },
          {
            "_id": "4fg53d85af965433a8b3a435",
            "appName": "Aunt Bertha",
            "appLogo": "https://url_to_app_logo_location",
            "appUrl": "https://url_to_app_url_location",
            "longDescription": "Long description of app",
            "shortDescription": "Short description of app",
            "isActive": true
          }
        ]
      },
...
      "UserApps": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "isActive": {
              "type": "boolean",
              "example": true
            },
            "consentedMedicalRecords": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "5ffc32eda3c2ec783e43b628"
              }
            },
            "_id": {
              "type": "string",
              "example": "608108b12cefd922906ae36d"
            },
            "app": {
              "type": "object",
              "properties": {
                "isDeleted": {
                  "type": "boolean",
                  "example": false
                },
                "isActive": {
                  "type": "boolean",
                  "example": true
                },
                "notifications": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "6006851ca3d8c04a9ce23f19,6006868fcb4b3041d0157393"
                  }
                },
                "medicalRecords": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "5fec6f918048d25b1c189a2e"
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "5fec6f6e8048d25b1c189a2d"
                },
                "appName": {
                  "type": "string",
                  "example": "Study App"
                },
                "longDescription": {
                  "type": "string",
                  "example": "Research patient case study"
                },
                "shortDescription": {
                  "type": "string",
                  "example": "Study App Short Description"
                },
                "createdBy": {
                  "type": "string",
                  "example": "Balwant"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2020-12-30T12:10:26.531Z"
                },
                "updatedBy": {
                  "type": "string",
                  "example": "username"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2020-12-30T12:10:26.531Z"
                },
                "appLogo": {
                  "type": "string",
                  "example": "studyapp_logo.svg"
                },
                "appUrl": {
                  "type": "string",
                  "example": ""
                },
                "consentVersion": {
                  "type": "string",
                  "example": 1
                },
                "consentTermOfUse": {
                  "type": "string",
                  "example": "test contents for consent terms of use"
                },
                "consentPrivacy": {
                  "type": "string",
                  "example": "test contents for consent privacy"
                },
                "consentPolicy": {
                  "type": "string",
                  "example": "<div>Consent Policy</div>"
                }
              }
            },
            "consentUpdatedAt": {
              "type": "string",
              "example": "2021-04-22T05:25:05.981Z"
            }
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}